Moved NAWS protocol header into detail namespace.
