+++ include/telnetpp/options/naws/detail/naws.hpp
@@ -2,8 +2,8 @@
 
 #include "telnetpp/core.hpp"
 
-namespace telnetpp { namespace options { namespace naws {
+namespace telnetpp { namespace options { namespace naws { namespace detail {
 
 static constexpr telnetpp::u8 option = 31;
 
-}}}
+}}}}
