+++ test/subnegotiation_router_test.cpp
@@ -1,6 +1,5 @@
 #include "telnetpp/detail/subnegotiation_router.hpp"
 #include "telnetpp/options/naws/client.hpp"
-#include "telnetpp/options/naws.hpp"
 #include "telnetpp/protocol.hpp"
 #include "telnetpp/detail/registration.hpp"
 #include <gtest/gtest.h>
@@ -8,32 +7,32 @@
 TEST(subnegotiation_router_test, when_nothing_is_registered_router_sinks_data)
 {
     telnetpp::detail::subnegotiation_router router;
-    router(telnetpp::subnegotiation(telnetpp::options::naws::option, {}));
+    router(telnetpp::subnegotiation(0, {}));
 }
 
 TEST(subnegotiation_router_test, message_with_registered_key_goes_to_registered_function)
 {
     telnetpp::detail::subnegotiation_router router;
-    
+
     telnetpp::subnegotiation sub(0x00, {});
     telnetpp::subnegotiation expected(0x01, {0x02});
     bool unregistered_route_called = false;
-    
-    router.register_route(expected.option(), 
+
+    router.register_route(expected.option(),
         [&sub](auto &&new_subnegotiation) -> std::vector<telnetpp::token>
-        { 
+        {
             sub = new_subnegotiation;
             return {};
         });
-    
+
     router.set_unregistered_route(
         [&unregistered_route_called](auto &&) -> std::vector<telnetpp::token>
         {
             unregistered_route_called = true;
             return {};
         });
-    
-    
+
+
     router(expected);
 
     ASSERT_EQ(expected, sub);
@@ -43,30 +42,30 @@ TEST(subnegotiation_router_test, message_with_registered_key_goes_to_registered_
 TEST(subnegotiation_router_test, message_with_unregistered_key_goes_to_unregistered_function)
 {
     telnetpp::detail::subnegotiation_router router;
-    
+
     telnetpp::subnegotiation sub(0x00, {});
     telnetpp::subnegotiation expected(0x01, {0x02});
     telnetpp::subnegotiation unexpected(0x02, {0x02});
-    
+
     bool registered_route_called = false;
-    
-    router.register_route(unexpected.option(), 
+
+    router.register_route(unexpected.option(),
         [&registered_route_called](auto &&) -> std::vector<telnetpp::token>
         {
             registered_route_called = true;
             return {};
         });
-    
+
     router.set_unregistered_route(
         [&sub](auto &&new_subnegotiation) -> std::vector<telnetpp::token>
         {
             sub = new_subnegotiation;
             return {};
         });
-    
-    
+
+
     router(expected);
-    
+
     ASSERT_EQ(false, registered_route_called);
     ASSERT_EQ(expected, sub);
 }
@@ -77,9 +76,9 @@ TEST(subnegotiation_router_test, routing_subnegotiation_returns_subnegotiation_r
     telnetpp::options::naws::client client;
     client.activate();
     client.negotiate(telnetpp::will);
-    
+
     telnetpp::detail::register_route_from_subnegotiation_to_option(router, client);
-    
+
     telnetpp::u16 width = 0;
     telnetpp::u16 height = 0;
     client.on_window_size_changed.connect(
@@ -90,14 +89,14 @@ TEST(subnegotiation_router_test, routing_subnegotiation_returns_subnegotiation_r
             height = new_height;
             return {};
         });
-    
+
     router(telnetpp::subnegotiation(
-        telnetpp::options::naws::option,
+        client.option(),
         { 0x00, 80, 0x00, 24 }));
-    
+
     telnetpp::u16 expected_width = 80;
     telnetpp::u16 expected_height = 24;
-    
+
     ASSERT_EQ(expected_width, width);
     ASSERT_EQ(expected_height, height);
 }
